[ui-storagebrowser] fixes css conflict due to antd module
